+++ src/main/java/com/braintreegateway/CustomActionsPaymentMethodDetails.java
@@ -10,12 +10,14 @@ public class CustomActionsPaymentMethodDetails {
   private String token;
   private String actionName;
   private String globalId;
+  private String uniqueNumberIdentifier;
   private List<CustomActionsPaymentMethodField> fields;
 
   public CustomActionsPaymentMethodDetails(NodeWrapper node) {
     token = node.findString("token");
     globalId = node.findString("global-id");
     actionName = node.findString("action-name");
+    uniqueNumberIdentifier = node.findString("unique-number-identifier");
     fields = new ArrayList<CustomActionsPaymentMethodField>();
     for (NodeWrapper fieldResponse : node.findAll("fields/field")) {
       fields.add(new CustomActionsPaymentMethodField(fieldResponse));
@@ -32,6 +34,10 @@ public class CustomActionsPaymentMethodDetails {
 
   public String getActionName() {
     return actionName;
+  }
+
+  public String getUniqueNumberIdentifier() {
+    return uniqueNumberIdentifier;
   }
 
   public List<CustomActionsPaymentMethodField> getFields() {