+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 2.102.1
 * Fix issue where billing address was not applied correctly in `ThreeDSecureLookupRequest`
+* Add `getUniqueNumberIdentifier` to `CustomActionsPaymentMethodDetails`
 
 ## 2.102.0
 * Add `networkResponseCode` and `networkResponseText` fields to `Transaction` and `CreditCardVerification`